requirements.txt
allure-behave==2.3.1b1
allure-python-commons==2.3.1b1
astroid==1.6.2
attrs==17.4.0
behave==1.2.5
enum34==1.1.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
parse==1.8.2
parse-type==0.4.2
Pillow==5.0.0
pkg-resources==0.0.0
pluggy==0.6.0
PyAutoGUI==0.9.36
PyHamcrest==1.9.0
pylint==1.8.3
PyMsgBox==1.0.6
PyMySQL==0.8.0
PyScreeze==0.1.14
python-xlib==0.21
PyTweening==1.0.3
selenium==3.8.1
six==1.11.0
wrapt==1.10.11
xlib==0.21
xlrd==1.1.0
